Done requested changes in KojiJenkinsTest, added test for fake-koji preview, created new FakeKojiTestUtil class for common code for fake-koji tests
