+++ fake-koji/pom.xml
@@ -47,6 +47,17 @@
                         </manifest>
                     </archive>
                 </configuration>
+                <!--
+                To be able to reuse test classes in different project
+                See: http://maven.apache.org/guides/mini/guide-attached-tests.html
+                -->
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
