+++ jenkins-scm-koji-plugin/pom.xml
@@ -56,6 +56,17 @@
             <version>0.1-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
+        <!--
+          Reuse test classes in different project
+          See: http://maven.apache.org/guides/mini/guide-attached-tests.html
+        -->
+        <dependency>
+            <groupId>fake-koji</groupId>
+            <artifactId>fake-koji</artifactId>
+            <version>0.1-SNAPSHOT</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <repositories>
